requirements.txt
# ============================================================
# DNS Attack Detection - ML/DL System
# Requirements File
# Python 3.11+
# ============================================================

# Core Data Science
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# Network / DNS Capture
scapy>=2.5.0
pyshark>=0.6.0

# DNS Utilities
dnspython>=2.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# API (optional FastAPI alerts)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
python-json-logger>=2.0.7
click>=8.1.7
psutil>=5.9.0

# Jupyter
notebook>=7.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0
